INTTEMPLATES-13 - Allow for unlimited nesting of top-level-package
Added for the following template:

* si-template-standalone-artifactId
* si-template-standalone-simple-artifactId

Template *si-template-war-artifactId* will be covered by *INTTEMPLATES-11*.
The *adapter* template will be fixed by *INTTEMPLATES-12*.
