+++ si-template-projects/si-template-standalone-project/build.gradle
@@ -24,7 +24,7 @@ eclipse {
 	}
 }
 
-mainClassName = "stsorg.stsspringframework.stsintegration.Main"
+mainClassName = "sipackage.Main"
 
 run{
