+++ si-template-projects/si-template-standalone-project/src/main/java/sipackage/Main.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package stsorg.stsspringframework.stsintegration;
+package sipackage;
 
 import java.util.Scanner;
