+++ si-template-projects/si-template-standalone-project/src/main/resources/log4j.xml
@@ -15,7 +15,7 @@
 		<level value="warn" />
 	</logger>
 
-	<logger name="stsorg.stsspringframework.stsintegration">
+	<logger name="sipackage">
 		<level value="info" />
 	</logger>
