+++ si-template-projects/si-template-standalone-project/wizard.json
@@ -30,14 +30,14 @@
 					"type" : "java.lang.String",
 					"page" : 0,
 					"required" : true,
-					"pattern" : "\\w+\\.\\w+(\\.\\w+)+",
+					"pattern" : "\\w+(\\.\\w+)*",
 					"replaceKind" : "topLevelPackage"
 				}
 			]
 		},
-			"topLevelPackage" : "stsorg.stsspringframework.stsintegration.mysubsystem",
-			"projectName" : "si-template-project-name",
-			"pages" : {
+		"topLevelPackage" : "sipackage",
+		"projectName" : "si-template-project-name",
+		"pages" : {
 			"page" : [
 				{
 					"order" : 0,
