+++ si-template-projects/si-template-standalone-simple-project/pom.xml
@@ -62,7 +62,7 @@
 				<artifactId>exec-maven-plugin</artifactId>
 				<version>1.2.1</version>
 				<configuration>
-					<mainClass>stsorg.stsspringframework.stsintegration.Main</mainClass>
+					<mainClass>sipackage.Main</mainClass>
 				</configuration>
 			</plugin>
 		</plugins>
